Refactor voyages into a hook
